# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /okx/ai/analyze:
    post:
      operationId: analyzeMarket
      tags: [ai]
      summary: AI analysis of the current market for an instrument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiAnalysisInput"
      responses:
        "200":
          description: AI-generated market analysis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiAnalysis"
  /okx/ai/recommend:
    post:
      operationId: recommendTrade
      tags: [ai]
      summary: Get parallel trade recommendations from multiple AI models
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiAnalysisInput"
      responses:
        "200":
          description: Recommendations from all configured models
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiRecommendationsResponse"
...
  /okx/ai/research:
    post:
      operationId: runResearchPipeline
      tags: [ai]
      summary: Full 3-stage research pipeline (technical agent + sentiment agent + 4-model decision battle)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiAnalysisInput"
      responses:
        "200":
          description: Pipeline result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchResult"
  /okx/auto/config:
    get:
      operationId: getAutoTradeConfig
      tags: [auto]
      summary: Get auto-trade configuration (singleton)
      responses:
        "200":
          description: Current config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutoTradeConfig"
    put:
      operationId: updateAutoTradeConfig
      tags: [auto]
      summary: Update auto-trade configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutoTradeConfigInput"
      responses:
        "200":
          description: Updated config
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutoTradeConfig"
...
  /okx/auto/decisions:
    get:
      operationId: listAutoTradeDecisions
      tags: [auto]
      summary: Recent AI decisions (auto + user-triggered)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Decisions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AiDecisionRecord"
  /okx/auto/executions:
    get:
      operationId: listAutoTradeExecutions
      tags: [auto]
      summary: Recent auto-trade executions (placed + closed orders)
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Executions list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AutoExecutionRecord"
...
components:
  schemas:
...
    AiAnalysisInput:
      type: object
      required: [instId]
      properties:
        instId:
          type: string
        mode:
          type: ["string", "null"]
          enum: [spot, perp, null]
          description: spot (default) or perp - changes the prompt for the AI
        marginUsdt:
          type: ["number", "null"]
          description: For perp mode, max USDT margin the AI may suggest committing
        maxLeverage:
          type: ["integer", "null"]
          description: For perp mode, max leverage the AI may suggest
    AiAnalysis:
      type: object
      required: [instId, analysis, generatedAt, model]
      properties:
        instId:
          type: string
        analysis:
          type: string
          description: Markdown-formatted analysis from the AI model
        generatedAt:
          type: string
        model:
          type: string
    AiRecommendation:
      type: object
      required: [providerId, providerLabel, model, latencyMs, ok]
      properties:
        providerId:
          type: string
          description: Stable id (anthropic | openai | gemini | openrouter)
        providerLabel:
          type: string
          description: Human-readable label (e.g. "Claude Sonnet 4.6")
        model:
          type: string
        latencyMs:
          type: integer
        ok:
          type: boolean
        error:
          type: ["string", "null"]
        action:
          type: ["string", "null"]
          enum: [buy, sell, hold, long, short, close, null]
        sizeUsdt:
          type: ["number", "null"]
          description: Spot mode - suggested USDT notional. Perp mode - leave null.
        marginUsdt:
          type: ["number", "null"]
          description: Perp mode - suggested USDT margin (notional = margin * leverage)
        leverage:
          type: ["integer", "null"]
          description: Perp mode - suggested leverage (1 to maxLeverage)
        stopLossPrice:
          type: ["number", "null"]
        takeProfitPrice:
          type: ["number", "null"]
          description: Perp mode - optional take-profit trigger price
        confidence:
          type: ["integer", "null"]
          description: 1-10 self-reported confidence
        reasoning:
          type: ["string", "null"]
    AiRecommendationsResponse:
      type: object
      required: [instId, generatedAt, lastPrice, recommendations]
      properties:
        instId:
          type: string
        generatedAt:
          type: string
        lastPrice:
          type: number
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/AiRecommendation"
    ResearchResult:
      type: object
      required: [instId, mode, generatedAt, lastPrice, recommendations]
      properties:
        instId:
          type: string
        mode:
          type: string
          enum: [spot, perp]
        generatedAt:
          type: string
        lastPrice:
          type: number
        technicalSummary:
          type: ["string", "null"]
        sentimentSummary:
          type: ["string", "null"]
        indicatorTextByBar:
          type: ["string", "null"]
        contextText:
          type: ["string", "null"]
        fundingRate:
          type: ["number", "null"]
        openInterestCcy:
          type: ["number", "null"]
        longShortRatio:
          type: ["number", "null"]
        takerBuyRatio:
          type: ["number", "null"]
        atr1H:
          type: ["number", "null"]
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/AiRecommendation"
    AutoTradeConfig:
      type: object
      required: [enabled, whitelist, maxMarginPctPerTrade, maxDailyLossPct, maxConcurrentPositions, maxLeverage, minConsensusCount, minAvgConfidence, cooldownMinutes, updatedAt]
      properties:
        enabled:
          type: boolean
        whitelist:
          type: array
          items:
            type: string
        maxMarginPctPerTrade:
          type: number
        maxDailyLossPct:
          type: number
        maxConcurrentPositions:
          type: integer
        maxLeverage:
          type: integer
        minConsensusCount:
          type: integer
        minAvgConfidence:
          type: integer
        cooldownMinutes:
          type: integer
        killUntil:
          type: ["string", "null"]
        updatedAt:
          type: string
    AutoTradeConfigInput:
      type: object
      properties:
        enabled:
          type: ["boolean", "null"]
        whitelist:
          type: ["array", "null"]
          items:
            type: string
        maxMarginPctPerTrade:
          type: ["number", "null"]
        maxDailyLossPct:
          type: ["number", "null"]
        maxConcurrentPositions:
          type: ["integer", "null"]
        maxLeverage:
          type: ["integer", "null"]
        minConsensusCount:
          type: ["integer", "null"]
        minAvgConfidence:
          type: ["integer", "null"]
        cooldownMinutes:
          type: ["integer", "null"]
...
    AiDecisionRecord:
      type: object
      required: [id, createdAt, instId, mode, lastPrice, recommendations, triggeredBy]
      properties:
        id:
          type: integer
        createdAt:
          type: string
        instId:
          type: string
        mode:
          type: string
        lastPrice:
          type: number
        technicalSummary:
          type: ["string", "null"]
        sentimentSummary:
          type: ["string", "null"]
        consensusAction:
          type: ["string", "null"]
        consensusConfidence:
          type: ["integer", "null"]
        triggeredBy:
          type: string
        recommendations:
          type: array
          items:
            $ref: "#/components/schemas/AiRecommendation"
    AutoExecutionRecord:
      type: object
      required: [id, createdAt, instId, side, status]
      properties:
        id:
          type: integer
        createdAt:
          type: string
        decisionId:
          type: ["integer", "null"]
        instId:
          type: string
        side:
          type: string
        marginUsdt:
          type: ["number", "null"]
        leverage:
          type: ["integer", "null"]
        contracts:
          type: ["number", "null"]
        entryPrice:
          type: ["number", "null"]
        ordId:
          type: ["string", "null"]
        status:
          type: string
        reason:
          type: ["string", "null"]
        realizedPnlUsdt:
          type: ["number", "null"]
        closePrice:
          type: ["number", "null"]
        closedAt:
          type: ["string", "null"]
        chosenProviderId:
          type: ["string", "null"]
...