# lib/api-zod/src/generated/types/aiDecisionRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiRecommendation } from "./aiRecommendation";

export interface AiDecisionRecord {
  id: number;
  createdAt: string;
  instId: string;
  mode: string;
  lastPrice: number;
  /** @nullable */
  technicalSummary?: string | null;
  /** @nullable */
  sentimentSummary?: string | null;
  /** @nullable */
  consensusAction?: string | null;
  /** @nullable */
  consensusConfidence?: number | null;
  triggeredBy: string;
  recommendations: AiRecommendation[];
}