# lib/api-zod/src/generated/types/autoExecutionRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AutoExecutionRecord {
  id: number;
  createdAt: string;
  /** @nullable */
  decisionId?: number | null;
  instId: string;
  side: string;
  /** @nullable */
  marginUsdt?: number | null;
  /** @nullable */
  leverage?: number | null;
  /** @nullable */
  contracts?: number | null;
  /** @nullable */
  entryPrice?: number | null;
  /** @nullable */
  ordId?: string | null;
  status: string;
  /** @nullable */
  reason?: string | null;
  /** @nullable */
  realizedPnlUsdt?: number | null;
  /** @nullable */
  closePrice?: number | null;
  /** @nullable */
  closedAt?: string | null;
  /** @nullable */
  chosenProviderId?: string | null;
}