# lib/api-zod/src/generated/types/autoTradeConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AutoTradeConfig {
  enabled: boolean;
  whitelist: string[];
  maxMarginPctPerTrade: number;
  maxDailyLossPct: number;
  maxConcurrentPositions: number;
  maxLeverage: number;
  minConsensusCount: number;
  minAvgConfidence: number;
  cooldownMinutes: number;
  /** @nullable */
  killUntil?: string | null;
  updatedAt: string;
}