# lib/api-zod/src/generated/types/autoTradeConfigInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AutoTradeConfigInput {
  /** @nullable */
  enabled?: boolean | null;
  /** @nullable */
  whitelist?: string[] | null;
  /** @nullable */
  maxMarginPctPerTrade?: number | null;
  /** @nullable */
  maxDailyLossPct?: number | null;
  /** @nullable */
  maxConcurrentPositions?: number | null;
  /** @nullable */
  maxLeverage?: number | null;
  /** @nullable */
  minConsensusCount?: number | null;
  /** @nullable */
  minAvgConfidence?: number | null;
  /** @nullable */
  cooldownMinutes?: number | null;
}