# lib/api-zod/src/generated/types/researchResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiRecommendation } from "./aiRecommendation";
import type { ResearchResultMode } from "./researchResultMode";

export interface ResearchResult {
  instId: string;
  mode: ResearchResultMode;
  generatedAt: string;
  lastPrice: number;
  /** @nullable */
  technicalSummary?: string | null;
  /** @nullable */
  sentimentSummary?: string | null;
  /** @nullable */
  indicatorTextByBar?: string | null;
  /** @nullable */
  contextText?: string | null;
  /** @nullable */
  fundingRate?: number | null;
  /** @nullable */
  openInterestCcy?: number | null;
  /** @nullable */
  longShortRatio?: number | null;
  /** @nullable */
  takerBuyRatio?: number | null;
  /** @nullable */
  atr1H?: number | null;
  recommendations: AiRecommendation[];
}